dadude/requirements.v2.txt
# DaDude v2.0 - Python Dependencies
# PostgreSQL + Redis Edition

# ===========================================
# Web Framework
# ===========================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# ===========================================
# Templates (legacy support)
# ===========================================
jinja2>=3.1.0

# ===========================================
# RouterOS API
# ===========================================
routeros-api>=0.17.0

# ===========================================
# Database - PostgreSQL
# ===========================================
sqlalchemy>=2.0.0
asyncpg>=0.29.0              # PostgreSQL async driver (replaces aiosqlite)
psycopg2-binary>=2.9.9       # PostgreSQL sync driver (for tooling/migrations)
alembic>=1.13.0              # Database migrations
greenlet>=3.0.0              # Required for SQLAlchemy async

# ===========================================
# Cache & Sessions - Redis
# ===========================================
redis>=5.0.0                 # Redis client
aioredis>=2.0.0              # Redis async (compatibility layer)

# ===========================================
# HTTP Client
# ===========================================
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# ===========================================
# Validation & Configuration
# ===========================================
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# ===========================================
# Utilities
# ===========================================
python-dateutil>=2.8.0
croniter>=2.0.0

# ===========================================
# Logging
# ===========================================
loguru>=0.7.0

# ===========================================
# MAC Vendor Lookup (OUI Database)
# ===========================================
mac-vendor-lookup>=0.1.12

# ===========================================
# Background Tasks
# ===========================================
apscheduler>=3.10.0

# ===========================================
# SNMP (optional)
# ===========================================
pysnmp>=4.4.0

# ===========================================
# WMI / Windows
# ===========================================
impacket>=0.11.0

# ===========================================
# SSH
# ===========================================
paramiko>=3.3.0

# ===========================================
# Encryption
# ===========================================
cryptography>=41.0.0

# ===========================================
# WebSocket (mTLS agent communication)
# ===========================================
websockets>=12.0

# ===========================================
# File handling
# ===========================================
python-multipart>=0.0.6

# ===========================================
# Progress bars (for migration scripts)
# ===========================================
tqdm>=4.66.0
